Поправлен фильтр
